Optimizacion y correccion de errores

+++ mobile/lib/models/estacion.dart
@@ -2,13 +2,13 @@ class Estacion {
   final int id;
   final String nombre;
   final String ubicacion;
-  final int lectura; // <--- AGREGAMOS ESTO
+  final double lectura; // double, no int — el backend devuelve Float
 
   Estacion({
-    required this.id, 
-    required this.nombre, 
+    required this.id,
+    required this.nombre,
     required this.ubicacion,
-    required this.lectura, // <--- TAMBIÉN AQUÍ
+    required this.lectura,
   });
 
   factory Estacion.fromJson(Map<String, dynamic> json) {
@@ -16,7 +16,7 @@ class Estacion {
       id: json['id'],
       nombre: json['nombre'],
       ubicacion: json['ubicacion'],
-      lectura: json['valor'] ?? 0,
+      lectura: (json['valor'] ?? 0).toDouble(), // conversión segura a double
     );
   }
 }
